+++ packages/docs/stories/mock-requests/index.js
@@ -8,8 +8,8 @@ import slotmachine from './slotmachine';
 import me from './me';
 import settings from './settings';
 import disclaimers from './disclaimers';
+import webQL from './webQL';
 import resourceSelect from './resource-select';
-import thanos from './thanos';
 
 mock.setup();
 
@@ -31,7 +31,7 @@ settings(mock);
 
 disclaimers(mock);
 
-thanos(mock);
+webQL(mock);
 
 resourceSelect(mock);