+++ packages/reactstrap-validation-date/typings/AvInput.d.ts
@@ -1,22 +1,21 @@
-export type AvInput = {
+export type LimitType = {
+    value: number;
+    units: number;
+};
+export interface AvInput extends React.HTMLAttributes<HTMLFormElement> {
     name: string;
     validationEvent?: '' | 'onInput' | 'onChange' | 'onBlur' | 'onFocus' | Array<'onInput' | 'onChange' | 'onBlur' | 'onFocus'>;
     validate?:object;
     value?: any;
     defaultValue?: any;
+    min?: string | LimitType;
+    max?: string | LimitType;
     trueValue?: any;
     falseValue?: any;
     checked?: boolean;
     defaultChecked?: boolean;
     state?: boolean;
-    type?: string;
     multiple?: boolean;
     disabled?: boolean;
     readOnly?: boolean;
-    onKeyUp?: Function;
-    onInput?: Function;
-    onFocus?: Function;
-    onBlur?: Function;
-    onChange?: Function;
-    onReset?: Function;
 }