+++ packages/page-header/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.11](https://github.com/Availity/availity-react/compare/@availity/page-header@4.4.10...@availity/page-header@4.4.11) (2019-07-31)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
 ## [4.4.10](https://github.com/Availity/availity-react/compare/@availity/page-header@4.4.9...@availity/page-header@4.4.10) (2019-07-23)
 
 **Note:** Version bump only for package @availity/page-header